Use the channel directly in methods
This simplifies the code as well as avoiding needless synchronisation
once the channel is ready. Because the methods really *do* have to
synchronise, I patch them initially with versions that do so, then
unpatch themselves.

Any setup that must be done before other operations (e.g.,
constructing a queue for replies) is done in the constructor, with the
final promise used as the 'ready' signal for unpatching.