Use confirms for flow control
The new-style streams API asks writable streams to invoke a callback
for each write, so that it can keep track of which are complete and
thereby limit its buffer. The channel API I'm writing to is similar to
the external API of streams (returns false if the buffer is 'full'). I
can either

 - just invoke the callback regardless of the write result; or,
 - start maintaining my own buffer; or,
 - use confirmations to signal successful writes

Of these, the first quickly leads to filling socket buffers, since the
stream implementation will happily keep writing (as I'm telling it
each write succeeds immediately). This tends to starve other I/O.

The second is just a pain, which seems contrary to the spirit of
newstreams; perhaps I've missed a vital, helpful bit of the API.

The last is rather heavy weight, since it means we wait for not only a
RTT but a RTT to a queue (thought this is mitigated a bit by RabbitMQ
coalescing acks). However it does also have the benefit of making all
writes that bit more certain, which serves the interest of
simplification.