Let REQs have >1 request in progress
They still have to reply (write) in the order that the requests came
in, so if answering requires any asynchrony, the application will
either have to do more work to reorder answers, or only read one at a
time as before.

I've removed the prefetch for the REQ and REP. In the case of REP,
it's no longer necessary to be entirely synchronous. For REQ, it
wasn't enforcing anything much anyway -- only that the reply would be
buffered before another could be sent.